[CLEAN CODE] Replace deprecated api by newer

+++ tmail-backend/mailbox/elasticsearch-v6/src/main/java/org/apache/james/mailbox/elasticsearch/search/ElasticSearchSearcher.java
@@ -122,7 +122,7 @@ public class ElasticSearchSearcher {
     }
 
     private int computeRequiredSize(Optional<Integer> limit) {
-        return limit.map(value -> Math.min(value.intValue(), size))
+        return limit.map(value -> Math.min(value, size))
             .orElse(size);
     }
